backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.1
langchain-openai==0.1.1
openai==1.10.0
pymupdf==1.23.7
faiss-cpu==1.7.4
websockets==12.0
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
tiktoken==0.5.2
numpy==1.24.3
pandas==2.1.3
httpx==0.25.0
scikit-learn==1.3.2
scipy==1.11.3
nltk==3.8.1
chromadb==0.4.18
tqdm==4.66.1
sentence-transformers==2.2.2